% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainnetome.R
\docType{data}
\name{brainnetome}
\alias{brainnetome}
\title{Brainnetome Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
data(brainnetome)
}
\description{
Brain atlas for the Brainnetome cortical parcellation with 105 regions
per hemisphere. Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
data(brainnetome)
brainnetome
}
\references{
Fan L, Li H, Zhuo J, Zhang Y, Wang J, Chen L, Yang Z,
Chu C, Xie S, Laird AR, Fox PT, Eickhoff SB, Yu C, Jiang T (2016).
The Human Brainnetome Atlas: A New Brain Atlas Based on Connectional
Architecture. \emph{Cerebral Cortex}, 26(8):3508-3526.
\doi{10.1093/cercor/bhw157}
}
\concept{ggseg_atlases}
\keyword{datasets}
