% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainnetome.R
\docType{data}
\name{brainnetome_3d}
\alias{brainnetome_3d}
\title{Brainnetome 3d atlas}
\format{
An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
brainnetome_3d
}
\description{
The functional specialization of the human cingulate cortex analyzed by the resting-state functional
 connectivity (FC) of each subregion from a network perspective.
}
\examples{
data(brainnetome_3d)
}
\references{
Yu C, Zhou Y, Liu Y, Jiang T, Dong H, Zhang Y, Walter M.
    Functional segregation of the human cingulate cortex is confirmed by functional connectivity based neuroanatomical parcellation.
    Neuroimage. 2011 Feb 14;54(4):2571-81.
    doi: 10.1016/j.neuroimage.2010.11.018.
    Epub 2010 Nov 10. PMID: 21073967.
    (\href{https://pubmed.ncbi.nlm.nih.gov/21073967/}{PubMed})
}
\concept{ggseg3d_atlases}
\keyword{datasets}
