% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainnetome.R
\name{brainnetome_sub}
\alias{brainnetome_sub}
\title{Brainnetome Subcortical Atlas}
\usage{
brainnetome_sub()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Brain atlas for the Brainnetome subcortical parcellation with 36
subregions covering amygdala, hippocampus, thalamus, caudate, putamen,
pallidum, and nucleus accumbens.
}
\examples{
brainnetome_sub()
\dontrun{plot(brainnetome_sub())}
}
\references{
Fan L, Li H, Zhuo J, Zhang Y, Wang J, Chen L, Yang Z,
Chu C, Xie S, Laird AR, Fox PT, Eickhoff SB, Yu C, Jiang T (2016).
The Human Brainnetome Atlas: A New Brain Atlas Based on Connectional
Architecture. \emph{Cerebral Cortex}, 26(8):3508-3526.
\doi{10.1093/cercor/bhw157}
}
\seealso{
Other ggseg_atlases: 
\code{\link{brainnetome}()}
}
\concept{ggseg_atlases}
\concept{subcortical_atlases}
