% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegBrainnetome-package.R
\docType{package}
\name{ggsegBrainnetome-package}
\alias{ggsegBrainnetome}
\alias{ggsegBrainnetome-package}
\title{ggsegBrainnetome: Brainnetome Atlas for the 'ggseg' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Brainnetome atlas for the 'ggseg' ecosystem. Provides a 'ggseg_atlas' object with 2D polygon geometry for use with 'ggseg'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggseg/ggsegBrainnetome}
  \item Report bugs at \url{https://github.com/ggseg/ggsegBrainnetome/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

}
\keyword{internal}
